Test passing the default resolver
