+++ strawberry/field.py
@@ -5,6 +5,7 @@ import sys
 from typing import (
     TYPE_CHECKING,
     Any,
+    Awaitable,
     Callable,
     Dict,
     List,
@@ -24,6 +25,7 @@ from strawberry.annotation import StrawberryAnnotation
 from strawberry.arguments import StrawberryArgument
 from strawberry.exceptions import InvalidDefaultFactoryError, InvalidFieldArgument
 from strawberry.type import StrawberryType, StrawberryTypeVar
+from strawberry.types.info import Info
 from strawberry.union import StrawberryUnion
 from strawberry.unset import UNSET
 
@@ -43,6 +45,7 @@ UNRESOLVED = object()
 
 class StrawberryField(dataclasses.Field):
     python_name: str
+    default_resolver: Callable[[Any, str], object] = getattr
 
     def __init__(
         self,
@@ -143,6 +146,19 @@ class StrawberryField(dataclasses.Field):
         self.base_resolver = resolver
 
         return self
+
+    def get_result(
+        self, source: Any, info: Info, args: List[Any], kwargs: Dict[str, Any]
+    ) -> Union[Awaitable[Any], Any]:
+        """
+        Calls the resolver defined for the StrawberryField. If the field doesn't have a
+        resolver defined we default to using getattr on `source`.
+        """
+
+        if self.base_resolver:
+            return self.base_resolver(*args, **kwargs)
+
+        return self.default_resolver(source, self.python_name)  # type: ignore
 
     @property
     def arguments(self) -> List[StrawberryArgument]:
