+++ strawberry/schema/schema_converter.py
@@ -483,15 +483,11 @@ class GraphQLCoreConverter:
                 source=_source, info=info, kwargs=kwargs
             )
 
-            if hasattr(field, "get_result"):
-                return field.get_result(
-                    _source, info=info, args=field_args, kwargs=field_kwargs
-                )
+            field.default_resolver = self.config.default_resolver  # type: ignore
 
-            if field.base_resolver:
-                return field.base_resolver(*field_args, **field_kwargs)
-
-            return self.config.default_resolver(_source, field.python_name)
+            return field.get_result(
+                _source, info=info, args=field_args, kwargs=field_kwargs
+            )
 
         def _resolver(_source: Any, info: GraphQLResolveInfo, **kwargs):
             strawberry_info = _strawberry_info_from_graphql(info)
